Move around the code responsible for decorating builtin diagnostics
